refactor(taillog): move log file listing endpoint to TailLogResource #35208
Addresses review feedback: co-locate log file listing next to the existing
tail endpoint so /v1/logs becomes a coherent collection resource.

  GET /v1/logs                      list available log files (moved)
  GET /v1/logs/{fileName}/_tail     tail one log file (existing)

The endpoint keeps its admin + Maintenance portlet auth — same security
perimeter as the JSP it replaces. This is stricter than the existing
_tail method on the same resource (back-end user only); the asymmetry is
intentional and noted in javadoc. _downloadLog and _downloadClusterLog
remain in MaintenanceResource, untouched by this change.
