+++ dotcms-integration/src/test/java/com/dotcms/rest/api/v1/maintenance/MaintenanceResourceIntegrationTest.java
@@ -19,7 +19,6 @@ import com.dotcms.mock.request.MockAttributeRequest;
 import com.dotcms.mock.request.MockHttpRequestIntegrationTest;
 import com.dotcms.mock.response.MockHttpResponse;
 import com.dotcms.rest.ResponseEntityJobStatusView;
-import com.dotcms.rest.ResponseEntityListStringView;
 import com.dotcms.rest.ResponseEntityStringView;
 import com.dotcms.rest.ResponseEntityView;
 import com.dotcms.rest.api.v1.job.JobStatusResponse;
@@ -31,12 +30,10 @@ import com.dotcms.util.IntegrationTestInitService;
 import com.dotmarketing.beans.Host;
 import com.dotmarketing.business.APILocator;
 import com.dotmarketing.portlets.contentlet.model.Contentlet;
-import com.dotmarketing.util.Config;
 import com.dotmarketing.util.UUIDGenerator;
 import com.liferay.portal.model.User;
 import com.liferay.portal.util.WebKeys;
 import java.io.File;
-import java.io.IOException;
 import java.nio.file.Files;
 import java.util.Arrays;
 import java.util.Collections;
@@ -198,70 +195,6 @@ public class MaintenanceResourceIntegrationTest extends IntegrationTestBase {
     public void test_deletePushedAssets_asNonAdmin_throwsSecurity() {
         final HttpServletRequest request = createRequestForUser(nonAdminUser);
         resource.deletePushedAssets(request, mockResponse);
-    }
-
-    // ==================== GET /_logFiles ====================
-
-    /**
-     * Given scenario: redirect {@code TAIL_LOG_LOG_FOLDER} to an isolated temp directory and
-     *                 plant two matching {@code .log} files and one non-matching {@code .txt}
-     *                 file inside it
-     * Expected result: both {@code .log} files appear in the returned list (sorted alphabetically
-     *                  and stripped of the base path), and the {@code .txt} file is filtered out
-     *
-     * <p>The override is fully restored in {@code finally} so the test never pollutes the real
-     * dotsecure/logs directory or leaks state across tests.
-     */
-    @Test
-    public void test_listLogFiles_filtersAndSortsAndStripsBasePath() throws Exception {
-        final String originalLogFolder = Config.getStringProperty(
-                "TAIL_LOG_LOG_FOLDER", "./dotsecure/logs/");
-
-        final File tempLogFolder = Files.createTempDirectory("maint-loglist-").toFile();
-        tempLogFolder.deleteOnExit();
-
-        final File firstMatch = plantLogFile(tempLogFolder, "aa-itlist.log");
-        final File secondMatch = plantLogFile(tempLogFolder, "zz-itlist.log");
-        final File notMatching = plantLogFile(tempLogFolder, "aa-itlist.txt");
-
-        try {
-            Config.setProperty("TAIL_LOG_LOG_FOLDER", tempLogFolder.getAbsolutePath());
-
-            final ResponseEntityListStringView result =
-                    resource.listLogFiles(createAdminRequest(), mockResponse);
-
-            assertNotNull(result);
-            final List<String> names = result.getEntity();
-            assertNotNull(names);
-
-            assertTrue("listing should contain the first matching .log file: " + names,
-                    names.contains(firstMatch.getName()));
-            assertTrue("listing should contain the second matching .log file: " + names,
-                    names.contains(secondMatch.getName()));
-            assertFalse("non-matching .txt file must be filtered out by TAIL_LOG_FILE_REGEX: " + names,
-                    names.contains(notMatching.getName()));
-
-            // base path is stripped — no absolute path leaks through, no leading separator
-            final String basePath = tempLogFolder.getAbsolutePath();
-            for (final String name : names) {
-                assertFalse("listed name must not contain the absolute log folder path: " + name,
-                        name.contains(basePath));
-                assertFalse("listed name must not start with a separator: " + name,
-                        name.startsWith(File.separator));
-            }
-
-            // sort order: aa- before zz- among our planted files
-            assertTrue("aa-* entry must sort before zz-* entry: " + names,
-                    names.indexOf(firstMatch.getName()) < names.indexOf(secondMatch.getName()));
-        } finally {
-            Config.setProperty("TAIL_LOG_LOG_FOLDER", originalLogFolder);
-            FileUtils.deleteQuietly(tempLogFolder);
-        }
-    }
-
-    @Test(expected = SecurityException.class)
-    public void test_listLogFiles_asNonAdmin_throwsSecurity() {
-        resource.listLogFiles(createRequestForUser(nonAdminUser), mockResponse);
     }
 
     // ==================== POST /assets/_fix ====================
@@ -640,13 +573,6 @@ public class MaintenanceResourceIntegrationTest extends IntegrationTestBase {
 
     private HttpServletRequest createAdminRequest() {
         return createRequestForUser(adminUser);
-    }
-
-    private static File plantLogFile(final File folder, final String name) throws IOException {
-        final File file = new File(folder, name);
-        Files.writeString(file.toPath(), "test content");
-        file.deleteOnExit();
-        return file;
     }
 
     private static HttpServletRequest createRequestForUser(final User user) {
