+++ dotcms-integration/src/test/java/com/dotcms/rest/api/v1/taillog/TailLogResourceTest.java
@@ -2,13 +2,25 @@ package com.dotcms.rest.api.v1.taillog;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
+import com.dotcms.datagen.TestUserUtils;
+import com.dotcms.datagen.UserDataGen;
+import com.dotcms.mock.request.MockAttributeRequest;
+import com.dotcms.mock.request.MockHttpRequestIntegrationTest;
+import com.dotcms.mock.response.MockHttpResponse;
+import com.dotcms.rest.ResponseEntityListStringView;
 import com.dotcms.rest.api.v1.taillog.TailLogResource.MyTailerListener;
 import com.dotcms.rest.api.v1.taillog.TailLogResource.MyTailerThread;
+import com.dotcms.rest.exception.SecurityException;
 import com.dotcms.util.IntegrationTestInitService;
+import com.dotmarketing.business.APILocator;
+import com.dotmarketing.util.Config;
 import com.dotmarketing.util.Logger;
+import com.liferay.portal.model.User;
+import com.liferay.portal.util.WebKeys;
 import io.vavr.Tuple;
 import io.vavr.Tuple2;
 import java.io.BufferedWriter;
@@ -25,6 +37,9 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import org.apache.commons.io.FileUtils;
 import org.glassfish.jersey.media.sse.EventOutput;
 import org.glassfish.jersey.media.sse.OutboundEvent;
 import org.junit.Assert;
@@ -36,9 +51,22 @@ public class TailLogResourceTest {
     static final String TAILING = "Tailing info from file";
     static final String WRITTEN = "New Line written With Number";
 
+    private static TailLogResource resource;
+    private static HttpServletResponse mockResponse;
+    private static User adminUser;
+    private static User nonAdminUser;
+
     @BeforeClass
     public static void prepare() throws Exception {
         IntegrationTestInitService.getInstance().init();
+
+        resource = new TailLogResource();
+        mockResponse = new MockHttpResponse().response();
+        adminUser = TestUserUtils.getAdminUser();
+
+        nonAdminUser = new UserDataGen().nextPersisted();
+        APILocator.getRoleAPI().addRoleToUser(
+                APILocator.getRoleAPI().loadBackEndUserRole(), nonAdminUser);
     }
 
     synchronized void writeText(final Writer out, final int index) throws IOException {
@@ -135,6 +163,85 @@ public class TailLogResourceTest {
         assertEquals("keepAlive", name);
         final Boolean bool = (Boolean) data.get("keepAlive");
         assertTrue(bool);
+    }
+
+    // ==================== GET /v1/logs (listLogFiles) ====================
+
+    /**
+     * Given scenario: redirect {@code TAIL_LOG_LOG_FOLDER} to an isolated temp directory and
+     *                 plant two matching {@code .log} files and one non-matching {@code .txt}
+     *                 file inside it
+     * Expected result: both {@code .log} files appear in the returned list (sorted alphabetically
+     *                  and stripped of the base path), and the {@code .txt} file is filtered out
+     *
+     * <p>The override is fully restored in {@code finally} so the test never pollutes the real
+     * dotsecure/logs directory or leaks state across tests.
+     */
+    @Test
+    public void test_listLogFiles_filtersAndSortsAndStripsBasePath() throws Exception {
+        final String originalLogFolder = Config.getStringProperty(
+                "TAIL_LOG_LOG_FOLDER", "./dotsecure/logs/");
+
+        final File tempLogFolder = Files.createTempDirectory("taillog-list-").toFile();
+        tempLogFolder.deleteOnExit();
+
+        final File firstMatch = plantLogFile(tempLogFolder, "aa-itlist.log");
+        final File secondMatch = plantLogFile(tempLogFolder, "zz-itlist.log");
+        final File notMatching = plantLogFile(tempLogFolder, "aa-itlist.txt");
+
+        try {
+            Config.setProperty("TAIL_LOG_LOG_FOLDER", tempLogFolder.getAbsolutePath());
+
+            final ResponseEntityListStringView result =
+                    resource.listLogFiles(createRequestForUser(adminUser), mockResponse);
+
+            assertNotNull(result);
+            final List<String> names = result.getEntity();
+            assertNotNull(names);
+
+            assertTrue("listing should contain the first matching .log file: " + names,
+                    names.contains(firstMatch.getName()));
+            assertTrue("listing should contain the second matching .log file: " + names,
+                    names.contains(secondMatch.getName()));
+            assertFalse("non-matching .txt file must be filtered out by TAIL_LOG_FILE_REGEX: " + names,
+                    names.contains(notMatching.getName()));
+
+            // base path is stripped — no absolute path leaks through, no leading separator
+            final String basePath = tempLogFolder.getAbsolutePath();
+            for (final String name : names) {
+                assertFalse("listed name must not contain the absolute log folder path: " + name,
+                        name.contains(basePath));
+                assertFalse("listed name must not start with a separator: " + name,
+                        name.startsWith(File.separator));
+            }
+
+            // sort order: aa- before zz- among our planted files
+            assertTrue("aa-* entry must sort before zz-* entry: " + names,
+                    names.indexOf(firstMatch.getName()) < names.indexOf(secondMatch.getName()));
+        } finally {
+            Config.setProperty("TAIL_LOG_LOG_FOLDER", originalLogFolder);
+            FileUtils.deleteQuietly(tempLogFolder);
+        }
+    }
+
+    @Test(expected = SecurityException.class)
+    public void test_listLogFiles_asNonAdmin_throwsSecurity() {
+        resource.listLogFiles(createRequestForUser(nonAdminUser), mockResponse);
+    }
+
+    private static HttpServletRequest createRequestForUser(final User user) {
+        final HttpServletRequest request = new MockAttributeRequest(
+                new MockHttpRequestIntegrationTest("localhost", "/").request()
+        ).request();
+        request.setAttribute(WebKeys.USER, user);
+        return request;
+    }
+
+    private static File plantLogFile(final File folder, final String name) throws IOException {
+        final File file = new File(folder, name);
+        Files.writeString(file.toPath(), "test content");
+        file.deleteOnExit();
+        return file;
     }
 
 }
